Pre pull from likeButton
