+++ src/components/Publish/index.js
@@ -10,29 +10,18 @@ import { getTimelinePosts } from "../../services/linkrAPI";
 import PostsContext from "../../contexts/postsContext";
 import UserContext from "../../contexts/userContext";
 import { Container, LeftDiv, Form, Input, Button, RightDiv } from "./style";
+import checkHashtags from '../functions/checkHashtags';
 
 export default function Publish() {
   const [form, setForm] = React.useState({ link: "", text: "" });
   const [isLoading, setIsLoading] = React.useState(false);
-  const { setArrPosts, setArrTrendingHashtags } = React.useContext(PostsContext);
+  const { setArrPosts, setArrTrendingHashtags, refreshFeed, setRefreshFeed } = React.useContext(PostsContext);
   const { userData } = React.useContext(UserContext);
 
   function validateUrl(value) {
     return /^(?:(?:(?:https?|http):)?\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:[/?#]\S*)?$/i.test(
       value
     );
-  }
-
-  function checkHashtags(text) {
-    const splitText = text.split(" ");
-    let hashtags = [];
-
-    for (let i = 0; i < splitText.length; i++) {
-      if (splitText[i][0] === "#") {
-        hashtags.push(splitText[i].substring(1));
-      }
-    }
-    return hashtags;
   }
 
   function handleForm(e) {
@@ -107,6 +96,7 @@ export default function Publish() {
       });
 
     setIsLoading(false);
+    setRefreshFeed(!refreshFeed);
   }
 
   return (
